retreiving Categories from DB Category Class
