+++ src/classes/Category.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Store;
+
+class Category {
+    private $id;
+    private $categoryName;
+
+    //getters?
+
+    //setters?
+
+}
+
+
+//function displayCategories ($array):string {
+//
+//    $HTMLString = "";
+//    $path = "/src/app/categories.php?id=";
+//
+//    foreach ($array as $category){
+//        $HTMLString .= "<a href='" . $path . $category["id"] . "' >" . $category["categoryName"] . "</a><br>" ;
+//    }
+//
+//    return $HTMLString;
+//}
